# Zero-copy zero-allocation shared buffer between JVM and NumPy arrays