--- a/src/tests/JNIGridnetSharedMemClientSelfPlay.java
+++ b/src/tests/JNIGridnetSharedMemClientSelfPlay.java
@@ -33,6 +33,7 @@ import rts.UnitAction;
 import rts.UnitActionAssignment;
 import rts.units.Unit;
 import rts.units.UnitTypeTable;
+import util.NDBuffer;
 import weka.core.pmml.jaxbbindings.False;
 
 /**