This is the first part of the PR, another one will be done for Python project separately (submodule needs to be updated first).

PR introduces new set of `SharedMem` JNI clients that take `IntBuffer`s for obs, unit masks, and actions masks. Rather than returning those from function calls as native `int[]` arrays, clients update data in-place. `NDBuffer` is just a convenient wrapper to hide quite a lot of offsets calculations.

The idea is the following:
* CPython allocates `bytebuffer` with the capacity requires for storing entire `ndarray`
* buffer is converted into directly allocated `java.nio.IntBuffer` on JVM
* `np.asarray` is used to map the same buffer to NumPy array with further taking view onto the same memory with `.reshape` 

Now both CPython and JVM can exchange data between runtimes without any mem copy operations. Profiling before and after the change is done:

```
Before the change:

         5214440 function calls (5210344 primitive calls) in 38.373 seconds

   ncalls  tottime  percall  cumtime  percall filename:lineno(function)
     4096   14.419    0.004   32.318    0.008 vec_env.py:161(step_wait)
    98304    4.061    0.000    7.997    0.000 vec_env.py:136(_encode_obs)
     4096    2.126    0.001    5.373    0.001 vec_env.py:146(step_async)
   208994    9.132    0.000    9.132    0.000 {built-in method numpy.array}


After the change:

         1955869 function calls (1951773 primitive calls) in 22.445 seconds

   Ordered by: standard name

   ncalls  tottime  percall  cumtime  percall filename:lineno(function)
     4096   15.620    0.004   15.980    0.004 vec_env2.py:184(step_wait)

Masks, before the change:

         992307 function calls (990003 primitive calls) in 12.858 seconds

   Ordered by: standard name

   ncalls  tottime  percall  cumtime  percall filename:lineno(function)
      768    0.281    0.000    5.841    0.008 vec_env.py:196(get_action_mask)
    39940    7.300    0.000    7.300    0.000 {built-in method numpy.array}

Masks, after the change:

         378698 function calls (376394 primitive calls) in 5.340 seconds

   Ordered by: standard name

   ncalls  tottime  percall  cumtime  percall filename:lineno(function)
      768    0.878    0.001    0.878    0.001 vec_env2.py:250(get_action_mask)
     1538    0.069    0.000    0.069    0.000 {built-in method numpy.array}
```

Both changes together yields 2.5x speedup (even larger when running more clients within same `VecClient` instance).

Note, that memory allocation is done on CPython side because CPython runtime lives longer than JVM's. So the allocation has to be done this way to avoid mem corruption when JVM finalizer are called on JVM shutdown.