@kachayev 2022-01-12
Pushed optional thread pool to execute `gameStep` and `getMasks`. So far it gives 2x speedup with 8 threads. I assume it will be even better if we have shared buffer for actions rather than passing `int[]` between threads.

Without changes:

```
         2048969 function calls (2036681 primitive calls) in 29.740 seconds

   Ordered by: standard name

   ncalls  tottime  percall  cumtime  percall filename:lineno(function)
     4096   17.710    0.004   18.082    0.004 vec_env.py:425(step_wait)
     4096    0.051    0.000    0.051    0.000 vec_env.py:428(<listcomp>)
     4096    6.013    0.001    6.013    0.001 vec_env.py:431(get_action_mask)
```

With 8 threads:

```
         2051597 function calls (2039309 primitive calls) in 17.368 seconds

   Ordered by: standard name

   ncalls  tottime  percall  cumtime  percall filename:lineno(function)
     4096    9.196    0.002    9.558    0.002 vec_env.py:425(step_wait)
     4096    0.054    0.000    0.054    0.000 vec_env.py:428(<listcomp>)
     4096    1.955    0.000    1.955    0.000 vec_env.py:431(get_action_mask)
```