@kachayev 2022-01-14
Added shared buffer for actions using `numpy.copyto`. Gets us another 25% improvements:

```
         208716 function calls (196428 primitive calls) in 13.708 seconds

   Ordered by: standard name
```